backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.22
openai>=1.50.0
anthropic>=0.40.0
openpyxl>=3.1.5
python-dotenv>=1.0.1
pydantic-settings>=2.6.0
aiosmtplib>=3.0.0
python-multipart>=0.0.12
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
psycopg2-binary>=2.9.9
slowapi>=0.1.9
